requirements.txt
mcp
fastapi
uvicorn
pandas
requests
google-genai
NseKit